agent/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
apscheduler==3.10.4
yfinance==0.2.40
supabase==2.4.0
httpx==0.27.0
feedparser==6.0.11
google-genai>=0.8.0
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b
python-dotenv==1.0.1
